feat(insurance-needs): finished refactor split into sub components
