+++ src/components/clients/insurance-needs/index.ts
@@ -1,4 +1,7 @@
 export { InsuranceNeedsCard } from "./insurance-needs-card";
+export { InsuranceNeedsInputPanel } from "./insurance-needs-input-panel";
+export { InsuranceNeedsResults } from "./insurance-needs-results";
+export { InsuranceNeedsTrace } from "./insurance-needs-trace";
 export { InsuranceNeedsChart } from "./insurance-needs-chart";
 // Client-only chart component for dynamic import usage
 export { InsuranceNeedsChartClient } from "./insurance-needs-chart-client";
